Extract Hip904DisabledTest and Hip904FeatureFlagTest
